test: create test for review model

+++ db/models/review.js
@@ -2,7 +2,7 @@ module.exports = db => db.define('reviews', {
   title: STRING,
   content: TEXT,
   num_stars: {
-    type: INTEGER,
+    type: FLOAT,
     validate: {
       min: 1, max: 5
     }
